Remove compiler warning on OpenBSD because `time_t` is `long long` there
This warning was generated in OpenBSD 7.7:

```
cmds.c:584:24: warning: format specifies type 'long' but the argument
has type 'time_t' (aka 'long long') [-Wformat]
                        name, dur, dur == 1 ? "" : "s");
                              ^~~
1 warning generated.
```

time_t is `long long` on OpenBSD, but `long` on Linux and various other
systems. Use `long long` for `dur` in bench_dist_func() when using
fprintf() to make it work everywhere without getting a compiler warning.

d84b2128-3f25-11f0-b7df-83850402c3ce

+++ src/cmds.c
@@ -580,8 +580,14 @@ static int bench_dist_func(const char *name,
 	assert(fnc);
 	assert(br);
 
-	fprintf(stderr, "Looping %s() for %ld second%s...",
-	                name, dur, dur == 1 ? "" : "s");
+	/*
+	 * time_t is `long long` on OpenBSD, but `long` on Linux and various 
+	 * other systems. Use `long long` for `dur` to make it work everywhere 
+	 * without generating a compiler warning.
+	 */
+	fprintf(stderr, "Looping %s() for %lld second%s...",
+	                name, (long long)dur, dur == 1 ? "" : "s");
+
 	fflush(stderr);
 
 	br->name = name;
